Bug Fixes
Fixed an uh oh in adding the base game fish to the chances and making them spawn at 100% quantity. All fix come out of the water at 100% now as well as all modded and base game fish are catchable. Also added the smaller fish to be catchable with the large fish trap.